grpc: Fix 32-bit target APIs (#537)

+++ protobuf/protobuf-core/src/nativeMain/kotlin/kotlinx/rpc/protobuf/internal/WireEncoder.native.kt
@@ -65,7 +65,7 @@ internal class WireEncoderNative(private val sink: Sink) : WireEncoder {
      */
     private inner class Ctx {
         fun write(buf: CPointer<ByteVar>, size: Int): Boolean {
-            sink.writeFully(buf, 0L, size.toLong())
+            sink.writeFully(buf, 0L, size)
             return true
         }
     }
